BC-12022 - Refactor imports in error mapper and loggable interfaces; update circular dependency check command
