+++ src/infra/logger/interfaces/loggable.ts
@@ -1,4 +1,4 @@
-import { type ErrorLogMessage, type LogMessage, type ValidationErrorLogMessage } from '../interfaces';
+import { type ErrorLogMessage, type LogMessage, type ValidationErrorLogMessage } from './logging.interface';
 
 export interface Loggable {
 	getLogMessage(): LogMessage | ErrorLogMessage | ValidationErrorLogMessage;
